Moved sadmd object into a class file
